refacto: smart flush 책임 클래스로 분리
